Fix some pkg problems after pkg restructure
