+++ export/distro/init.go
@@ -10,13 +10,11 @@ import (
 	"fmt"
 	"github.com/edgexfoundry/edgex-go/internal"
 	"github.com/edgexfoundry/edgex-go/internal/pkg/consul"
+	"github.com/edgexfoundry/edgex-go/pkg/clients/coredata"
+	"github.com/edgexfoundry/edgex-go/pkg/clients/types"
 	"go.uber.org/zap"
 	"strconv"
 	"strings"
-
-	"github.com/edgexfoundry/edgex-go/core/clients/coredata"
-	"github.com/edgexfoundry/edgex-go/core/clients/types"
-	"github.com/edgexfoundry/edgex-go/internal"
 )
 
 const (
